Added support for excluded files

+++ Program.cs
@@ -1,15 +1,25 @@
 ﻿using System;
 using System.IO;
 using System.ServiceProcess;
-using System.Threading.Tasks;
 using System.Threading;
+using System.Linq;
+using System.Security.Principal;
 
 namespace Updater
 {
 	public class Program
 	{
+
 		static void Main(string[] args)
 		{
+			//Asks for admin privilege if we don't have it 
+			if (!IsAdministrator())
+			{
+				Console.WriteLine("This program must be running with administrator rights");
+				Thread.Sleep(5000);
+				Environment.Exit(0);
+			}
+
 			Config[] configs = Config.LoadConfig("Config.json"); //Load Configurations
 
 			int monitoredFolders = configs.Length;
@@ -47,52 +57,59 @@ namespace Updater
 				}
 				void Filewatcher_Created(object sender, FileSystemEventArgs e)
 				{
-					Log($" [{currentConfig.ServiceName}] - {e.Name} Detected. Waiting for Changes");
+					Log($"[{currentConfig.ServiceName}] - {e.Name} Detected. Waiting for Changes");
 					Thread.Sleep(15 * 1000); //15 seg
 
 					using ServiceController serviceController = new(currentConfig.ServiceName);
 					ServiceControllerStatus serviceStatus = serviceController.Status;
 
-					//Paramos el servicio
-					if (serviceStatus == ServiceControllerStatus.Running)
-					{
-						Log($" [{currentConfig.ServiceName}] - Stopping \"{currentConfig.ServiceName}\"");
-						serviceController.Stop();
-					} else
-					{
-						Log($" [{currentConfig.ServiceName}] - \"{currentConfig.ServiceName}\" is not running");
-					}
 
-					//Movemos Archivos
 					try
 					{
-						Log($" [{currentConfig.ServiceName}] - Deleting \"{currentConfig.ProgramFolder}\"");
+						//Paramos el servicio
+						if (serviceStatus == ServiceControllerStatus.Running)
+						{
+							Log(currentConfig, $"Stopping Service");
+							serviceController.Stop();
+						} else
+						{
+							Log(currentConfig, $"Service is not running");
+						}
+
+						//Movemos Archivos
+						Log(currentConfig, $"Deleting \"{currentConfig.ProgramFolder}\"");
 						Thread.Sleep(1 * 1000);
-						EmptyFolder(currentConfig.ProgramFolder);
+						EmptyFolder(currentConfig.ProgramFolder, currentConfig);
 
-						Log($" [{currentConfig.ServiceName}] - Copying \"{currentConfig.SupervisedFolder}\" to \"{currentConfig.ProgramFolder}\"");
-						CopyFilesRecursively(currentConfig.SupervisedFolder, currentConfig.ProgramFolder);
+						Log(currentConfig, $"Copying \"{currentConfig.SupervisedFolder}\" to \"{currentConfig.ProgramFolder}\"");
+						CopyFilesRecursively(currentConfig.SupervisedFolder, currentConfig.ProgramFolder, currentConfig);
 						Thread.Sleep(1 * 1000);
 
-						Log($" [{currentConfig.ServiceName}] - Deleting \"{currentConfig.SupervisedFolder}\"");
-						EmptyFolder(currentConfig.SupervisedFolder);
+						Log(currentConfig, $"Deleting \"{currentConfig.SupervisedFolder}\"");
+						EmptyFolder(currentConfig.SupervisedFolder, currentConfig, true);
 
-						Log($" [{currentConfig.ServiceName}] - Starting \"{currentConfig.ServiceName}\"");
+						Log(currentConfig, $" Starting Service");
 						serviceController.Start();
-						Log($" [{currentConfig.ServiceName}] - Service {currentConfig.ServiceName}. Updated Sucessfully");
+						Log(currentConfig, $"Service updated sucessfully");
 					}
 					catch (Exception ex)
 					{
-						Log($" [{currentConfig.ServiceName}] - Exception ocurred while updating" +
+						Log(currentConfig, $"Exception ocurred while updating" +
 						$"\n\t\t{ex.Message}");
 					}
 
-					static void EmptyFolder(string path)
+					static void EmptyFolder(string path, Config config, bool skipProtectedFiles = false)
 					{
 						DirectoryInfo directoryInfo = new(path);
 						foreach (FileInfo file in directoryInfo.GetFiles())
 						{
-							file.Delete();
+							if (!config.ProtectedFiles.Contains(Path.GetFileName(file.Name)) || skipProtectedFiles)
+							{
+								file.Delete();
+							} else
+							{
+								Log(config, $"Skipped deletion of {Path.GetFileName(file.Name)}.");
+							}
 						}
 						foreach (DirectoryInfo folder in directoryInfo.GetDirectories())
 						{
@@ -100,7 +117,7 @@ namespace Updater
 						}
 					}
 
-					static void CopyFilesRecursively(string sourcePath, string targetPath)
+					static void CopyFilesRecursively(string sourcePath, string targetPath, Config config)
 					{
 						//Now Create all of the directories
 						foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
@@ -111,7 +128,13 @@ namespace Updater
 						//Copy all the files & Replaces any files with the same name
 						foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
 						{
-							File.Copy(newPath, newPath.Replace(sourcePath, targetPath), true);
+							//Prevents overwriting the protected files
+							if (!config.ProtectedFiles.Contains(Path.GetFileName(newPath)))
+							{
+								File.Copy(newPath, newPath.Replace(sourcePath, targetPath), true);
+							} else {
+								Log(config, $"Skipped protected file {Path.GetFileName(newPath)}.");
+							}
 						}
 					}
 				}
@@ -124,7 +147,18 @@ namespace Updater
 
 		public static void Log(string output)
 		{
-			Console.WriteLine($"[{DateTime.Now:s}] - {output}");
+			Console.WriteLine($"{DateTime.Now:s} - {output}");
+		}
+
+		public static void Log(Config config,string output)
+		{
+			Console.WriteLine($"{DateTime.Now:s} - [{config.ServiceName}] - {output}");
+		}
+
+		public static bool IsAdministrator()
+		{
+			return (new WindowsPrincipal(WindowsIdentity.GetCurrent()))
+					  .IsInRole(WindowsBuiltInRole.Administrator);
 		}
 	}
 }
